Fix missing change from last commit

+++ bluetooth.hpp
@@ -363,7 +363,7 @@ class Bluetooth : public Stream
         unsigned long findBaud();
 
         bool isConnected();
-        bool waitForConnection(int timeout);
+        bool waitForConnection(long timeout);
         void printClientMAC(bool new_line = false);
         bool handlNewConnection();
